# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  entities.Class:
    properties:
      auditory:
        type: string
      date:
        type: string
      group_names:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
      teacher_names:
        items:
          type: string
        type: array
      time_end:
        type: string
      time_start:
        type: string
      type:
        type: string
      university:
        type: string
      week:
        type: integer
      weekday:
        type: integer
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  entities.ParseScheduleResponse:
    properties:
      message:
        type: string
    type: object
  entities.ScheduleGroups:
    properties:
      groups:
        items:
          type: string
        type: array
    type: object
  entities.ScheduleNames:
    properties:
      names:
        items:
          type: string
        type: array
    type: object
  entities.ScheduleTeachers:
    properties:
      teachers:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  title: VUZ+ API
  version: "1.0"
paths:
...
  /auth/schedule/group/{group}:
    get:
      consumes:
      - application/json
      parameters:
      - description: schedule group
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Class'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get schedule
      tags:
      - schedule
  /auth/schedule/group_subjects:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ScheduleNames'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search names
      tags:
      - schedule
...
  /auth/schedule/parse:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Upload schedule file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.ParseScheduleResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get schedule by name
      tags:
      - schedule
  /auth/schedule/search/group:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ScheduleGroups'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search groups
      tags:
      - schedule
  /auth/schedule/search/name:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ScheduleNames'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search names
      tags:
      - schedule
  /auth/schedule/search/teacher:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ScheduleTeachers'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Search teachers
      tags:
      - schedule
  /auth/schedule/teacher/{teacher}:
    get:
      consumes:
      - application/json
      parameters:
      - description: schedule teacher
        in: path
        name: teacher
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Class'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get schedule by teacher
      tags:
      - schedule
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"